+++ src/math/vec2.cc
@@ -4,7 +4,7 @@
 
 #include "math/vec2.h"
 
-vec2::vec2(double x, double y): x{x}, y{y} {}
+vec2::vec2(float x, float y): x{x}, y{y} {}
 
 vec2::vec2(const vec2& other): x{other.x}, y{other.y} {}
 vec2::vec2(vec2&& other): x{std::move(other.x)}, y{std::move(other.y)} {}
@@ -17,14 +17,14 @@ vec2& vec2::operator=(const vec2& other) {
 }
 
 vec2& vec2::normalize() {
-  double len = length();
+  float len = length();
   x /= len;
   y /= len;
 
   return *this;
 }
 
-vec2& vec2::scale(double s) {
+vec2& vec2::scale(float s) {
   x *= s;
   y *= s;
 
@@ -36,7 +36,7 @@ std::ostream& operator<<(std::ostream &os, const vec2& vec) {
   return os;
 }
 
-double vec2::length() const {
+float vec2::length() const {
   return sqrt(x * x + y * y);
 }
 
@@ -48,29 +48,29 @@ vec2 vec2::sub(const vec2& left, const vec2& right) {
   return left - right;
 }
 
-vec2 vec2::lerp(const vec2& start, const vec2& end, double t) {
+vec2 vec2::lerp(const vec2& start, const vec2& end, float t) {
   vec2 diff = end - start;
   return start + diff * t;
 }
 
-double vec2::dot(const vec2& left, const vec2& right) {
+float vec2::dot(const vec2& left, const vec2& right) {
   return left.x * right.x + left.y * right.y;
 }
 
-double vec2::angle(const vec2& left, const vec2& right) {
-  double cosA = dot(left, right) / (left.length() * right.length());
+float vec2::angle(const vec2& left, const vec2& right) {
+  float cosA = dot(left, right) / (left.length() * right.length());
   return acos(cosA);
 }
 
-double vec2::dist(const vec2& left, const vec2& right) {
-  double dx = right.x - left.x;
-  double dy = right.y - left.y;
+float vec2::dist(const vec2& left, const vec2& right) {
+  float dx = right.x - left.x;
+  float dy = right.y - left.y;
 
   return sqrt(dx * dx + dy * dy);
 }
 
 vec2 vec2::getNormalized(const vec2 &vec) {
-  double len = vec.length();
+  float len = vec.length();
 
   vec2 res;
   res.x = vec.x / len;
@@ -111,7 +111,7 @@ vec2 vec2::operator-(const vec2& right) const {
   return res;
 }
 
-vec2 vec2::operator*(double s) const {
+vec2 vec2::operator*(float s) const {
   vec2 res;
   res.x = x * s;
   res.y = y * s;
